delete comment backend
