+++ backend/app/app.ts
@@ -1,22 +1,22 @@
 import * as cors from 'cors'
 import * as bodyParser from 'body-parser'
 
-import PostController from './controllers/postRoutes'
+import PostRoutes from './routes/postRoutes'
 import serverConstructor from './serverConstructor'
-import AuthorizationController from './controllers/authorizationController'
-import CommentController from './controllers/commentController'
-import ProfileController from './controllers/profileRoutes'
+import AuthorizationRoutes from './routes/authorizationRoutes'
+import CommentRoutes from './routes/commentRoutes'
+import ProfileRoutes from './routes/profileRoutes'
 const cookieParser = require('cookie-parser')
 
 const { FRONTEND_URL } = process.env
 
 const backend = new serverConstructor({
     port: 8000,
-    controllers: [
-        new PostController(),
-        new AuthorizationController(),
-        new CommentController(),
-        new ProfileController()
+    routes: [
+        new PostRoutes(),
+        new AuthorizationRoutes(),
+        new CommentRoutes(),
+        new ProfileRoutes()
     ],
     middleWares: [
         cookieParser(),
