+++ backend/app/controllers/commentController.ts
@@ -1,21 +1,51 @@
-import ControllerBase from '../interfaces/ControllerBase'
-import * as express from 'express'
-import CommentService from '../service/commentService'
-const isAuth = require('../middleware/isAuth')
+import { Request, Response } from 'express'
+import CommentDAO from '../dao/commentDAO'
+import comment from '../entity/comment'
+import ProfileDAO from '../dao/profileDao'
 
-class CommentController implements ControllerBase{
-    public comment = '/comment'
-    public router = express.Router()
-    private commentService: CommentService
+class CommentController {
+    private dao: CommentDAO
+    private profileDAO: ProfileDAO
 
     constructor() {
-        this.commentService = new CommentService()
-        this.initRoutes()
+        this.dao = new CommentDAO()
+        this.profileDAO = new ProfileDAO()
     }
 
-    public initRoutes(): void {
-        this.router.get(this.comment + '/:path', this.commentService.getComments)
-        this.router.post(this.comment, isAuth, this.commentService.createComment)
+    public getComments = async (req: Request, res: Response): Promise<Response> => {
+        const foundComments = await this.dao.getComments(req.params.path)
+
+        if (req.params.path && foundComments)
+            return res.status(200).json(foundComments);
+        else
+            return res.status(404).json({ 'message': 'No comments found on that path' })
+    }
+
+    public createComment = async (req: Request | any, res: Response): Promise<Response> => {
+        const newComment = new comment()
+        const { path, content, parent_comment_id } = req.body
+        const profile = await this.profileDAO.getProfileByUsername(req.decoded.username)
+
+        newComment.profile = profile
+        newComment.path = path
+        newComment.content = content
+        newComment.parent_comment_id = parent_comment_id
+        newComment.created_at = new Date()
+        newComment.updated_at = new Date()
+
+        await this.dao.createComment(newComment)
+
+        return res.status(200).json({ message: 'Comment added!' })
+    }
+
+    public deleteComment = async (req: Request | any, res: Response): Promise<Response> => {
+        const comment = await this.dao.getComment(req.params.id)
+        if(comment.profile.display_name === req.decoded.username) {
+            await this.dao.archiveComment(req.params.id)
+            res.status(200).json({ message: 'Comment removed' })
+        }
+        return res.status(403).json({ message: 'Unauthorized' })
+
     }
 }
