+++ backend/app/routes/profileRoutes.ts
@@ -1,5 +1,5 @@
 import * as express from 'express'
-import ProfileService from '../service/profileService'
+import ProfileController from '../controllers/profileController'
 import RoutesBase from '../interfaces/ControllerBase'
 
 const auth = require('../middleware/isAuth.ts')
@@ -8,17 +8,17 @@ class ProfileRoutes implements RoutesBase {
     public profile = '/profile/:username?'
     public register = '/profile/register'
     public router = express.Router()
-    private profileService: ProfileService
+    private profileController: ProfileController
 
     constructor(){
-        this.profileService = new ProfileService()
+        this.profileController = new ProfileController()
         this.initRoutes()
     }
 
     public initRoutes(): void {
-        this.router.post(this.register, this.profileService.register)
-        this.router.put(this.profile, auth, this.profileService.updateProfile)
-        this.router.get (this.profile, this.profileService.getProfile)
+        this.router.post(this.register, this.profileController.register)
+        this.router.put(this.profile, auth, this.profileController.updateProfile)
+        this.router.get (this.profile, this.profileController.getProfile)
     }
 }
 export default ProfileRoutes
